+++ encoder/basisu_gpu_texture.cpp
@@ -15,7 +15,6 @@
 #include "basisu_gpu_texture.h"
 #include "basisu_enc.h"
 #include "basisu_pvrtc1_4.h"
-#include "3rdparty/android_astc_decomp.h"
 #include "basisu_bc7enc.h"
 #include "../transcoder/basisu_astc_hdr_core.h"
